Moved Pending Model to member.py, Renamed membership.py to member.py
