+++ squads/templates/squads/groups/view_group.html
@@ -18,7 +18,7 @@
     {% if not pending_application and not membership %}
         <form method="post" action="{% url 'squads:apply_group' group_id=group.id %}" class="d-flex align-items-end" style="gap: 0.5rem;">
             {% csrf_token %}
-            {% if group.require_approval %}
+            {% if group.req_approve %}
                 <div class="mr-2">
                     {{ comment_form.comment }}
                 </div>
@@ -26,9 +26,9 @@
             <input type="hidden" name="filter_req" value="{{ filter_req }}">
                 <button
                     type="submit"
-                    name="{% if group.require_approval %}apply_group{% else %}join_group{% endif %}"
-                    class="btn {% if group.require_approval %}btn-primary{% else %}btn-success{% endif %} {% if not filter_req %}disabled btn-danger{% endif %}">
-                    {% if group.require_approval %}
+                    name="{% if group.req_approve %}apply_group{% else %}join_group{% endif %}"
+                    class="btn {% if group.req_approve %}btn-primary{% else %}btn-success{% endif %} {% if not filter_req %}disabled btn-danger{% endif %}">
+                    {% if group.req_approve %}
                         {% translate "Apply" %}
                     {% else %}
                         {% translate "Join" %}
