+++ squads/templates/squads/manage/manage_members.html
@@ -24,6 +24,7 @@
                 <th class="col-user" style="width: 20%;">User</th>
                 <th class="col-group" style="width: 20%;">Group</th>
                 <th class="col-active" style="width: 20%;">Active</th>
+                <th class="col-filter" style="width: 20%;">Filters Met</th>
                 <th class="col-joined" style="width: 20%;">Joined</th>
                 <th class="col-actions" style="width: 10%;">Actions</th>
             </thead>
@@ -74,11 +75,19 @@ document.addEventListener('DOMContentLoaded', function () {
                     return '<button class="btn btn-danger btn-sm" disabled><i class="fa-solid fa-ban"></i></button>';
                 }
             }},
+            {data: 'req_filters', render: function(data, type, row) {
+                if (data) {
+                    return '<button class="btn btn-success btn-sm" disabled><i class="fa-solid fa-check"></i></button>';
+                } else {
+                    return '<button class="btn btn-danger btn-sm" disabled><i class="fa-solid fa-ban"></i></button>';
+                }
+            }},
             {data: 'joined_at'},
         ],
         columnDefs: [
+            {targets: [2,3], sortable: false},
             {
-                targets: 4,
+                targets: 5,
                 className: 'text-end',
                 sortable: false,
                 render: function (data, type, row) {
